+++ src/main/java/dnd/jackpot/project/repository/ProjectRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import dnd.jackpot.project.entity.Project;
 
-public interface ProjectRepository extends JpaRepository<Project, Integer>{
+public interface ProjectRepository extends JpaRepository<Project, Long>{
 //	프로젝트가 잘 업로드 및 불러오나 보기위해 일단 
 	Page<Project> findAll(Pageable pageable);
+	
 }
